fix: No public IP for ec2 instances
The instances created for terraform are in private network and doesn't require a public IP. Outbound connections are via NAT gateway and the command execution is via SSM which doesn't require any inbound connection.

Note: Technically this might not cause any issue. But it is unnecessary, confusing and not a good practice.